修复: 用户反馈功能的架构问题
- 重构 feedback 路由使用共享数据库连接而非创建新连接
- 在 db.ts 中新增 getUserFeedbackStats/List/ById 函数
- 添加 view_audit_log 权限检查保护反馈数据
- 清理 feishu.ts 中的敏感日志输出（移除完整事件数据）
- 删除不相关的 TOPIC_REACTION 文档（未实现功能）
- 删除临时修复脚本 fix-user-feedback-table.js
- 移动 check-feedback.js 到 scripts/dev/ 目录
